COM-2693 - fix reviews
