+++ src/types/UserType.ts
@@ -13,7 +13,7 @@ export type UserType = {
 export type AuxiliaryType = {
   _id: UserType['_id'],
   identity: UserType['identity'],
-  contact? : { phone?: string },
+  contact? : UserType['contact'],
   picture?: UserType['picture'],
   contracts?: UserType['contracts'],
   administrative?: UserType['administrative'],
